% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_info.R
\name{bibliolatamInfoUI}
\alias{bibliolatamInfoUI}
\title{Bibliolatam Info UI Component for Biblioshiny}
\usage{
bibliolatamInfoUI()
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
Renders the information and documentation tab for Bibliolatam inside biblioshiny.
}
\keyword{internal}
