% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_merge.R
\name{bibliolatamMergeServer}
\alias{bibliolatamMergeServer}
\title{Bibliolatam Merge Server Logic for Biblioshiny}
\usage{
bibliolatamMergeServer(input, output, session, values)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}

\item{values}{ReactiveValues global state.}
}
\description{
Bibliolatam Merge Server Logic for Biblioshiny
}
\keyword{internal}
