% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_biblioshiny.R
\name{export_biblioshiny}
\alias{export_biblioshiny}
\title{Export a converted collection to biblioshiny-compatible RData}
\usage{
export_biblioshiny(data, file)
}
\arguments{
\item{data}{An object of class \code{bibliometrixDB} or a compatible data frame.}

\item{file}{Path to the output \code{.RData} file.}
}
\value{
Invisibly returns the normalized path of the saved file.
}
\description{
Saves the bibliometric data frame as an \code{.RData} file ready to be loaded
directly in the \code{biblioshiny} web interface under 'Data -> Load bibliometrix RData'.
}
