% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_oasisbr.R
\name{oasisbrServer}
\alias{oasisbrServer}
\title{Oasisbr Server Logic for Biblioshiny}
\usage{
oasisbrServer(input, output, session, values)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{values}{ReactiveValues object from biblioshiny containing global state.}
}
\description{
Connects Oasisbr UI events with package parsers and the biblioshiny reactive state.
}
\keyword{internal}
