% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_oasisbr.R
\name{oasisbrUI}
\alias{oasisbrUI}
\title{Oasisbr UI Component for Biblioshiny}
\usage{
oasisbrUI()
}
\value{
A Shiny tagList containing the UI elements.
}
\description{
Renders the Oasisbr (IBICT) open-access collection panel
compatible with the biblioshiny AdminLTE layout.
}
\keyword{internal}
