% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_omnisearch.R
\name{omnisearchUI}
\alias{omnisearchUI}
\title{Omnisearch UI Component for Biblioshiny}
\usage{
omnisearchUI()
}
\value{
A Shiny tagList containing UI elements.
}
\description{
Renders the Latin American scientific Omnisearch panel compatible with biblioshiny.
}
\keyword{internal}
