% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bdtd.R
\name{read_bdtd}
\alias{read_bdtd}
\alias{read_oasisbr}
\title{Parse export files from BDTD / Oasisbr (IBICT)}
\usage{
read_bdtd(file, convert = TRUE)

read_oasisbr(file, convert = TRUE)
}
\arguments{
\item{file}{Path to the CSV or TSV file exported from BDTD/Oasisbr.}

\item{convert}{Logical. If TRUE, directly wraps output with \code{as_bibliometrix()}. Default is TRUE.}
}
\value{
A data frame formatted with standard bibliometric tags.
}
\description{
Reads and standardizes tabular exports (CSV, TSV) from the Biblioteca Digital
Brasileira de Teses e Dissertações (BDTD) or Oasisbr portal (IBICT).
}
