% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scielo.R
\name{read_scielo}
\alias{read_scielo}
\title{Parse export files from SciELO}
\usage{
read_scielo(file, convert = TRUE)
}
\arguments{
\item{file}{Path to the SciELO CSV export file.}

\item{convert}{Logical. If TRUE, directly wraps output with \code{as_bibliometrix(..., dbsource = "scielo")}. Default is TRUE.}
}
\value{
A data frame formatted with standard bibliometric tags.
}
\description{
Reads and standardizes tabular CSV export files downloaded from the SciELO portal (search.scielo.org).
}
