% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spell.R
\name{read_spell}
\alias{read_spell}
\title{Parse export files from SPELL / ANPAD}
\usage{
read_spell(file, convert = TRUE)
}
\arguments{
\item{file}{Path to the CSV file exported from SPELL.}

\item{convert}{Logical. If TRUE, directly wraps output with \code{as_bibliometrix()}. Default is TRUE.}
}
\value{
A data frame formatted with standard bibliometric tags.
}
\description{
Parse export files from SPELL / ANPAD
}
