% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_redalyc.R
\name{redalycServer}
\alias{redalycServer}
\title{Redalyc Server Logic for Biblioshiny}
\usage{
redalycServer(input, output, session, values)
}
\arguments{
\item{input}{Shiny input.}

\item{output}{Shiny output.}

\item{session}{Shiny session.}

\item{values}{ReactiveValues global state.}
}
\description{
Redalyc Server Logic for Biblioshiny
}
\keyword{internal}
