% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_scielo.R
\name{scieloServer}
\alias{scieloServer}
\title{SciELO Server Logic for Biblioshiny}
\usage{
scieloServer(input, output, session, values)
}
\arguments{
\item{input}{Shiny input object.}

\item{output}{Shiny output object.}

\item{session}{Shiny session object.}

\item{values}{ReactiveValues object from biblioshiny containing global state.}
}
\description{
Connects SciELO UI events with package parsers and the biblioshiny reactive state.
}
\keyword{internal}
