% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_redalyc.R
\name{redalycUI}
\alias{redalycUI}
\title{Redalyc UI Component for Biblioshiny}
\usage{
redalycUI()
}
\value{
A Shiny tagList containing UI elements.
}
\description{
Renders the Redalyc / AmeliCA collection import panel compatible with the biblioshiny layout.
}
\keyword{internal}
